# Optimize and fix original file paths