--- a/components/tools/OmeroPy/test/integration/gatewaytest/test_fs.py
+++ b/components/tools/OmeroPy/test/integration/gatewaytest/test_fs.py
@@ -179,6 +179,11 @@ class TestFileset(object):
         for image in fileset_with_images.copyImages():
             assert image.getFileset() is not None
 
+    def testGetImportedImageFiles(self, gatewaywrapper, fileset_with_images):
+        for image in fileset_with_images.copyImages():
+            files = list(image.getImportedImageFiles())
+            assert len(files) == 4
+
 
 class TestArchivedOriginalFiles(object):
 
@@ -246,3 +251,9 @@ class TestArchivedOriginalFiles(object):
     def testGetFileset(self, gatewaywrapper, images_with_original_files):
         for image in images_with_original_files:
             assert image.getFileset() is None
+
+    def testGetImportedImageFiles(
+            self, gatewaywrapper, images_with_original_files):
+        for image in images_with_original_files:
+            files = list(image.getImportedImageFiles())
+            assert len(files) == 2