This PR includes various performance improvements to the stack required to extract original file paths from the server via the `/webgateway/original_file_paths/...` OMERO.web JSON endpoint. Without these or similar changes there is the very real scenario where requesting the original file paths could **crash** your server due to the sheer amount and complexity of data being processed.

The "real world" example dataset which spurred the need for the improvements has 4608 Images (384 Wells, 12 Fields) and 4608 OriginalFiles (one per field) associated with the same Fileset. Half that number is needed to run a server out of memory with a 1GB heap. Call tree is as follows:

```
original_file_paths()
\___ _ImageWrapper.getImportedImageFiles()
\___ _ImageWrapper.getFileset()
\___ _ImageWrapper.countFilesetFiles()
\___ BlitzGateway.getFilesetFilesInfo()
```

The first set of improvements are to the fileset and archive file count methods, `getFilesetFileInfo()` and `getArchivedFilesInfo()`, on the gateway. Before:

```
...
In [25]: print datetime.now()
2015-04-03 08:45:46.464612

In [26]: gw.getFilesetFilesInfo([66101])

Out[26]: {'count': 4608, 'size': 6564240074L}

In [27]: print datetime.now()
2015-04-03 08:45:49.193313
...
```

After (14ee8844e62c434d8c4b827735a3d10424abc1b4):

```
In [12]: print datetime.now()
2015-04-03 14:38:57.178159

In [13]: gw.getFilesetFilesInfo([66101])
Out[13]: {'count': 4608L, 'size': 6564240074L}

In [14]: print datetime.now()
2015-04-03 14:38:57.371813
```

~2.7seconds vs. ~190ms (~14x speedup)

The second set of improvements are to the original file load method `getImportedImageFiles()` on the image wrapper. Before, even with a 2GB heap the aforementioned dataset causes out of memory and GC overhead limit exceptions. This is likely due to the multiplicity of the queries performed and the complexity of the object graph. The number of objects loaded would be >20,000.

After (1356aff4c2f3f18728cce5a0cc67973248285777):

```
In [11]: print datetime.now()
2015-04-03 19:04:43.403131

In [12]: files = list(image.getImportedImageFiles())
print datetime.now()

In [13]: print datetime.now()
2015-04-03 19:04:49.463482
```

~6seconds.

These projection changes complete the speedup.

After (27ce963e1a7e8a97ac346c3d68982f44855a2bf4), run on `localhost`:

```
$ time curl "https://localhost/webgateway/original_file_paths/66101/?server=1&username=...&password=..." &> /dev/null

real	0m0.356s
user	0m0.024s
sys	0m0.037s
```

~350ms

Integration tests cover the affected code paths. /cc @ximenesuk 

To be reviewed by @will-moore and/or @aleksandra-tarkowska.

/cc @jballanc, @knabar, @emilroz, @jburel, @sbesson