@sbesson 2015-04-08
Quick fire test performed on a large  SPIM dataset (~3600 files) by asking the Web client to display the file paths:
- without this PR, the query is started and the server hangs as described (7GB heap), requiring a restart of the server
- with this PR, the same query on the same dataset with a server with similar memory settings returns within 100ms.