@aleksandra-tarkowska 2015-04-10
@chris-allan I am struggling with a large dataset. I managed to test:

I managed to test ```getFilesetFilesInfo```:
 - with Plate 
```
Merge user-3
conn.getFilesetFilesInfo([24001])
{'count': 123L, 'annotations': [{'ns': None, 'id': None}], 'fileset': True, 'size': 1008799953L}
```

Paths on server: ```user-3_9/2015-03/12/09-07-54.472/plate1_G11.pnl ...```

```
$ du /repositories/OMERO-5.1-merge-deploy/ManagedRepository/user-3_9/2015-03/12/09-07-54.472/
985408	/repositories/OMERO-5.1-merge-deploy/ManagedRepository/user-3_9/2015-03/12/09-07-54.472/
```
looks like values are different. I tested with:

```
size = 0
for f in img.getFileset().listFiles():
    size +=f.size
print size
1008799953
```
might be worth to look

 - with SVS:
```
Merge: user-1
conn.getFilesetFilesInfo([230854])
{'count': 1L, 'annotations': [{'ns': None, 'id': None}], 'fileset': True, 'size': 340130041L}
statinfo.st_size: 340130041
```