@mtbc 2015-04-13
(As an aside: I don't know if `DiskUsage` or `UsedFilesRequest` from http://downloads.openmicroscopy.org/omero/5.1.0/api/slice2html/omero/cmd.html are of any use for this kind of thing, or if their performance is adequate under comparable stresses, but we may want to review their design if they could make the gateway's life easier.)