@chris-allan 2015-04-13
I would guess that difference is because `du` is dividing by 1024, @aleksandra-tarkowska:

```
In [2]: 1008799953 / 1024
Out[2]: 985156
```

The additional size (`985408 - 985156 = 252`) is the on disk overhead of the directories in that hierarchy.

Can't comment on the user of those commands, @mtbc, other than to say that as I recall they were fundamentally operating on IDs. Sending 1000's, at worst tens of thousands, of IDs back to the server then for an `IN` query so you can get a list of paths is certainly not ideal.

/cc @joshmoore