@aleksandra-tarkowska 2015-04-13
```
$ du --bytes /repositories/OMERO-5.1-merge-deploy/ManagedRepository/user-3_9/2015-03/12/09-07-54.472/
1008804049    /repositories/OMERO-5.1-merge-deploy/ManagedRepository/user-3_9/2015-03/12/09-07-54.472/

1008804049 - 1008799953 = 4096
```
thank @chris-allan for point that out