@will-moore 2015-04-17
I haven't repeated the performance testing above, but code changes seem fine and web functionality and UI appears unchanged. Good to merge.