@will-moore 2015-04-17
Batch download of a variety of files (with and without filesets) is also still working fine (uses the methods above).