@chris-allan 2015-04-17
Just realised there might be an issue if there are no annotations present based on the output of @aleksandra-tarkowska. I'll look at this today.